refactor(codegen): extrai calls/builtins.rs
Move 5 fns de builtins de tipo (~1239 LOC) para
expressions/calls/builtins.rs:
- lower_string_builtin / lower_number_builtin / lower_map_set_builtin
- lower_console_call / lower_array_builtin

emit_user_fn_addr em calls/mod.rs virou pub(crate) para uso
de builtins::lower_array_builtin (passes user fn refs).

calls/mod.rs: 4481 -> 3247 linhas. Suite verde.
